+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release
+* Generate Labels as per width of screen
+* Generate Labels as per steps given
+* Gradient Color for Container
+* Border Radius 
+
+